Init custom branch out of kickstart.nvim
Addition of:

- xiyaowong/transparent.nvim plugino
- clangd, clang-format, cpplint mason lsp servers

Note: Caveat on windows for clangd and format, cpplint.

Uninstall python from store if it was installed that way, install through `choco install python` and make sure `AppData\Local\Microsoft\WindowsApps` in path does not mess up with mason python discovery, if you have a lingering Python.exe file there. Move it down the list of regsitered path if need be.